+++ docs/src/dev-docs/changelog.rst
@@ -24,6 +24,24 @@ changelog <https://keepachangelog.com/en/1.1.0/>`_ format. This project follows
 Unreleased
 ----------
 
+Version 2025.8 - 2025-06-11
+---------------------------
+
+Changed
+#######
+
+- Checkpoints for fine-tuning files are now passed from the ``options.yaml``.
+
+Version 2025.7 - 2025-05-27
+---------------------------
+
+Changed
+#######
+
+- Metatrain is now built on top of ``metatomic`` instead of
+  ``metatensor.torch.atomistic``. Please refer to https://docs.metatensor.org/metatomic/
+  to find how to use the new models.
+
 Version 2025.6 - 2025-04-28
 ---------------------------